"Add count method to SampleService and SampleController
Implemented a method in SampleService to retrieve the count of SampleEntities in the repository. This count method was also exposed via a new GET mapping in SampleController to allow users to access this information directly through a new endpoint. This expansion improves the functionality of our application and enhances user experience by providing useful statistics."